requirements.txt
streamLit
torch
torchvision
numpy
pillow
opencv-python-headless
scikit-image
monai